# backend/src/main/kotlin/io/github/projektalmanac/comunidades/generated/dto/IdUsuarioDto.kt
package io.github.projektalmanac.comunidades.generated.dto

import java.util.Objects
import com.fasterxml.jackson.annotation.JsonProperty
import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Email
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size
import javax.validation.Valid
import io.swagger.v3.oas.annotations.media.Schema

/**
 * 
 * @param idUsuario 
 */
data class IdUsuarioDto(

    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("idUsuario", required = true) val idUsuario: kotlin.Int
) {

}