# backend/src/main/kotlin/io/github/projektalmanac/comunidades/generated/dto/UsuarioCreadoDto.kt
package io.github.projektalmanac.comunidades.generated.dto

import java.util.Objects
import com.fasterxml.jackson.annotation.JsonProperty
import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Email
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size
import javax.validation.Valid
import io.swagger.v3.oas.annotations.media.Schema

/**
 * 
 * @param id 
 * @param nombre 
 * @param apellidos 
 * @param correo 
 */
data class UsuarioCreadoDto(

    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("id", required = true) val id: kotlin.Int,

    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("nombre", required = true) val nombre: kotlin.String,

    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("apellidos", required = true) val apellidos: kotlin.String,

    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("correo", required = true) val correo: kotlin.String
) {

}